+++ src/ilp_solver_interface_t.cpp
@@ -37,7 +37,7 @@ namespace ILPSolver
 
         // Add variables
         for (auto i = 0; i < num_vars; ++i)
-            p_solver->add_variable_integer(1, 0, std::numeric_limits<double>::max(), "x" + std::to_string(i));
+            p_solver->add_variable_integer(1, 0, std::numeric_limits<int>::max(), "x" + std::to_string(i));
 
         // Add constraints
         cout << "Initial array: ";
